KEY MAPPING FUCKING WOOOORKS
But only for one player so far, often mupen64plus tells me that
there is a config input problem. For some reason, when the server runs
startx, a mupen64plus.cfg is already generated, and I (think, not sure)
it is overriding what is there, so that might be part of the problem.
But the mapping works for one player so who cares!